+++ src/run.rs
@@ -387,6 +387,7 @@ fn connect(
         request_level,
         now: timings.now(),
         timeout: timings.next_timeout(Timeout::Connect),
+        current_time: timings.current_time().clone(),
         run_connector: agent.run_connector.clone(),
     };